refactor: move metadata component into documents group
