+++ frontend/src/components/documents/Metadata.vue
